refactor: extract execution modules and consolidate constants
- Create CorpConstants.ts with shared constants (SPAWN_COOLDOWN, creep bodies, limits)
- Extract main.ts into focused modules:
  - execution/CorpRunner.ts: corp lifecycle management
  - execution/IncrementalAnalysis.ts: terrain analysis
  - execution/Persistence.ts: state persistence
  - execution/Visualization.ts: RoomVisual rendering
- main.ts reduced from 1,239 to 331 lines (73% reduction)
- Add corp serialization types to Memory.ts
